+++ app/Http/Controllers/UserController.php
@@ -12,6 +12,12 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
+        /*
+        if(!$request->user()->can("listar_user")){
+            abort(403);
+        }
+        */
+        $this->authorize("index_user");
         // /api/user?limit=5&page=2&q=juan
         $limit = isset($request->limit)?$request->limit:10;
         $buscar = isset($request->q)?$request->q:null;
@@ -32,6 +38,8 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize("store_user");
+
         $request->validate([
             "name" => "required",
             "email" => "required|email|unique:users",
@@ -52,6 +60,8 @@ class UserController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize("show_user");
+
         $usuario = User::findOrFail($id);
         return response()->json($usuario, 200);
     }
@@ -61,6 +71,9 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id)
     {
+
+        $this->authorize("update_user");
+
         $request->validate([
             "name" => "required",
             "email" => "required|email|unique:users,email,$id"
@@ -82,6 +95,8 @@ class UserController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize("delete_user");
+
         $usuario = User::findOrFail($id);
         $usuario->delete();
